% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cohort.R
\name{create_cohort}
\alias{create_cohort}
\title{Generate the base cohort}
\usage{
create_cohort(source_n, n)
}
\arguments{
\item{source_n}{number of EHR sources (sites).}

\item{n}{number of patients per EHR source.}
}
\value{
A tibble with 2 columns, pid (patient id) +/- source.
}
\description{
`create_cohort()` generates a tibble with the specified number of patients
from a specified number of sources (e.g., for multi-site databases).
}
\examples{
create_cohort(3, 150)
}
