% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_var_char.R
\name{gen_var_char}
\alias{gen_var_char}
\title{Generate a character variable}
\usage{
gen_var_char(v, nc, opts, p)
}
\arguments{
\item{v}{name of variable.}

\item{nc}{name of cohort used to match number of observations.}

\item{opts}{vector of options to be sampled from.}

\item{p}{vector of probabilities for opts. Must be same length as opts.}
}
\value{
a data.frame with # observations equal to the cohort.
}
\description{
`gen_var_char()` generates a character variable.
}
\examples{
gen_var_char(v = sex, nc = 100, opts = c("F","M"), p = c(0.51, 0.49))
}
