% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_var_num.R
\name{gen_var_num}
\alias{gen_var_num}
\title{Generate a character variable}
\usage{
gen_var_num(v, nc, dist, int = FALSE, ...)
}
\arguments{
\item{v}{name of variable. vector of options to be sampled from.}

\item{nc}{positive integer indicating number of observations to generate; or,
single data.frame from which to match observations}

\item{dist}{lowercase name of distribution (in quotes). Options are: normal}

\item{int}{logical indicating whether returned values should be rounded to
integers or not}

\item{...}{Other options used to control the random number generation, e.g., mean,
var, lamba, min, max, n, etc.}
}
\value{
a data.frame with # observations equal to `nc` dataset.
}
\description{
`gen_var_num()` generates a character variable.
}
\examples{
x <- gen_var_num(v = systolic, nc = 1000, dist = "normal", int = TRUE, mean = 130, sd = 16)
}
