% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdate.R
\name{rdate}
\alias{rdate}
\title{Generate random date(s)}
\usage{
rdate(
  v = rdate,
  size,
  min = paste0(format(Sys.Date(), "\%Y"), "-01-01"),
  max = Sys.Date()
)
}
\arguments{
\item{v}{name of variable.}

\item{size}{number of name of data.table to match size of.}

\item{min}{minimum date of date range to sample from, formatted as
"YYYY-MM-DD"; defaults to Jan 1 of current year.}

\item{max}{maximum date of date range to sample from, formatted as
"YYYY-MM-DD"; defaults to execution date.}
}
\value{
A date.
}
\description{
`rdate()` is an internal function that generates a random date, similarly
to other random number generators in base R.
}
\examples{
rdate(10)
rdate(10, "2010-12-01", "2015-11-30", sort = FALSE)
}
\seealso{
[add_date_var()], which is the user function this function supports.
}
