% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtime.R
\name{rtime}
\alias{rtime}
\title{Generate random time(s)}
\usage{
rtime(v = time, nc, format = "12H", pr.wh = 0.5, min = "00:00", max = "23:59")
}
\arguments{
\item{v}{name of variable.}

\item{nc}{name of data.table to match size of.}

\item{format}{final formatting of time; valid options are "12H" and "24H"}

\item{pr.wh}{probability of time being in 'working hours' (defined as 8 AM to
5 PM); a value of 0.5 means equal probability for working vs. non-working
hours. Primarily used to preferentially select working-hour admit or
measurement times for outpatient/ambulatory visits.}

\item{min}{earliest time to sample from (in 24H format); default is "00:00"}

\item{max}{latest time to sample from (in 24H format); default is "23:59"}
}
\value{
A dataset with character column formatted as "HH:MM" (24H) or
"HH:MM AM" or "HH:MM PM" (12H).
}
\description{
`rtime()` is an user-facing function that generates a random time, similarly
to other random number generators in base R.
}
\examples{
rdate(10)
rdate(10, "2010-12-01", "2015-11-30", sort = FALSE)
}
