% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_rtime.R
\name{update_rtime}
\alias{update_rtime}
\title{Generate random date(s)}
\usage{
update_rtime(perm_time, perm_date, fix_time, fix_date)
}
\arguments{
\item{perm_time}{variable that represents the established time around which
`fix_time` will be corrected.}

\item{perm_date}{variable that represents the established date around which
`fix_time` will be corrected.}

\item{fix_time}{variable that represents the to-be-updated time which will be
corrected.}

\item{fix_date}{variable that represents the date of the to-be-updated time
which will be corrected.}
}
\description{
`update_rtime()` is an user-facing function that updates (fixes) a previously-
generated time which does not logically make sense. Primarily used to update
randomly-generated discharge times, based on admit times/dates (e.g., when
the randomly-generated discharge time occurs before the randomly generated
admit time, and admit and discharge date are equivalent).
}
\examples{
encounters <- encounters |>
  dplyr::mutate(discharge_time = update_rtime(perm_time = admit_time,
                                              perm_date = admit_date,
                                              fix_time = discharge_time,
                                              fix_date = discharge_date))
#' @export
}
