% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_rtime.R
\name{update_rtime}
\alias{update_rtime}
\title{Generate random date(s)}
\usage{
update_rtime(data, perm_time, perm_date, fix_time, fix_date)
}
\arguments{
\item{data}{data.frame which needs updating.}

\item{perm_time}{existing time around which `fix_time` will be corrected.}

\item{perm_date}{the existing date around which `fix_time` will be corrected.}

\item{fix_time}{the to-be-corrected time.}

\item{fix_date}{the date of the to-be-corrected time.}
}
\description{
`update_rtime()` is an user-facing function that updates (fixes) a previously-
generated time which does not logically make sense. Primarily used to update
randomly-generated discharge times, based on admit times/dates (e.g., when
the randomly-generated discharge time occurs before the randomly generated
admit time, and admit and discharge date are equivalent).
}
\examples{
df <- data.frame(admit_date = c("2022-02-01", "2022-02-01"),
                 discharge_date = c("2022-02-01", "2022-02-03"),
                 admit_time = c("08:00", "10:00"),
                 discharge_time = c("06:00", "06:00"))

df <- update_rtime(data = df, perm_time = admit_time, perm_date = admit_date,
                   fix_time = discharge_time, fix_date = discharge_date)
}
